Extracted CoordTransformOptions.
